+++ gui/view_color_plane.go
@@ -109,7 +109,7 @@ func (pv *colorPlaneView) GenerateLayout(w *Window) Layout {
 			}),
 		},
 	}
-	colorControlFocusRing(&cv.cfg)
+	colorControlFocusRing(&cv.cfg, w.EffID(cv.cfg.ID))
 	return generateViewLayout(cv, w)
 }